got rid of eadk and shared. now everything is sorted correctly
